Json cleanup
