+++ kifi-backend/modules/common/app/com/keepit/common/json/package.scala
@@ -5,21 +5,23 @@ import play.api.data.validation.ValidationError
 
 package object json {
 
-  /* Inspired from https://coderwall.com/p/orci8g */
-  implicit def tuple2Reads[A, B](implicit aReads: Reads[A], bReads: Reads[B]): Reads[Tuple2[A, B]] = Reads[Tuple2[A, B]] {
-    case JsArray(arr) if arr.size == 2 => for {
-      a <- aReads.reads(arr(0))
-      b <- bReads.reads(arr(1))
-    } yield (a, b)
-    case _ => JsError(Seq(JsPath() -> Seq(ValidationError("Expected array of two elements"))))
-  }
+  object TupleFormat {
+    /* Inspired from https://coderwall.com/p/orci8g */
+    implicit def tuple2Reads[A, B](implicit aReads: Reads[A], bReads: Reads[B]): Reads[(A, B)] = Reads[(A, B)] {
+      case JsArray(arr) if arr.size == 2 => for {
+        a <- aReads.reads(arr(0))
+        b <- bReads.reads(arr(1))
+      } yield (a, b)
+      case _ => JsError(Seq(JsPath() -> Seq(ValidationError("Expected array of two elements"))))
+    }
 
-  implicit def tuple2Writes[A, B](implicit aWrites: Writes[A], bWrites: Writes[B]): Writes[Tuple2[A, B]] = new Writes[Tuple2[A, B]] {
-    def writes(tuple: Tuple2[A, B]) = JsArray(Seq(aWrites.writes(tuple._1), bWrites.writes(tuple._2)))
-  }
+    implicit def tuple2Writes[A, B](implicit aWrites: Writes[A], bWrites: Writes[B]): Writes[(A, B)] = new Writes[(A, B)] {
+      def writes(tuple: (A, B)) = JsArray(Seq(aWrites.writes(tuple._1), bWrites.writes(tuple._2)))
+    }
 
-  implicit def tuple2Format[A, B](implicit reads: Reads[(A, B)], writes: Writes[(A, B)]) = {
-    Format(reads, writes)
+    implicit def tuple2Format[A, B](implicit aReads: Format[A], bReads: Format[B]): Format[(A, B)] = {
+      Format(tuple2Reads[A, B], tuple2Writes[A, B])
+    }
   }
 
   implicit class PimpedFormat[A](format: Format[A]) {
@@ -38,4 +40,21 @@ package object json {
   }
 
   def minify(fullJson: JsObject): JsObject = JsObject(fullJson.fields.filterNot { case (key, value) => canBeOmitted(value) })
+
+  /* Be careful, this will always attempt to deserialize to Left first, and then to Right if Left has failed.
+  This could lead to an incorrect behavior if the Right type can be serialized to a valid Left type
+  So we're not providing this Format as an implicit on purpose. */
+
+  object EitherFormat {
+    def apply[A, B](implicit leftFormat: Format[A], rightFormat: Format[B]) = new Format[Either[A, B]] {
+      def reads(json: JsValue) = json.validate[A].map(Left(_)) orElse json.validate[B].map(Right(_))
+      def writes(either: Either[A, B]) = either.left.map(Json.toJson(_)).right.map(Json.toJson(_)).merge
+    }
+  }
+
+  object TraversableFormat {
+    private def materialize[U, T <: Traversable[U]](implicit traversableFormatter: Format[T]) = traversableFormatter
+    def seq[U](implicit formatter: Format[U]) = materialize[U, Seq[U]]
+    def set[U](implicit formatter: Format[U]) = materialize[U, Set[U]]
+  }
 }
