+++ kifi-backend/modules/common/app/com/keepit/search/augmentation/AugmentableItem.scala
@@ -3,7 +3,7 @@ package com.keepit.search.augmentation
 import com.keepit.common.db.{ ExternalId, Id }
 import com.keepit.model._
 import play.api.libs.json._
-import com.keepit.serializer.TupleFormat
+import com.keepit.common.json.TupleFormat
 
 case class AugmentableItem(uri: Id[NormalizedURI], keptIn: Option[Id[Library]] = None)
